refactor: extract play page reducer and scoring logic
## Changes

### 1. Extracted reducer to separate module
- **Created**: `web/src/features/quiz/playReducer.ts` (141 lines)
- Moved all state management logic out of component
- Exported `PlayState`, `PlayAction`, and constants
- Added `createInitialState()` factory function
- **Benefit**: Reducer can now be unit tested independently

### 2. Consolidated scoring logic
- **Created**: `web/src/lib/scoring.ts` (56 lines)
- Moved `computePoints()` - calculates score based on time
- Moved `rollupTally()` - updates running score
- Moved `composeSummary()` - creates result summary
- Added comprehensive JSDoc comments
- **Benefit**: Scoring logic is now testable and reusable

### 3. Simplified play page component
- **Updated**: `web/app/play/page.tsx`
- Reduced from 463 to 422 lines (-41 lines, -9%)
- Removed 170+ lines of inline definitions
- Cleaner imports with focused responsibilities
- **Benefit**: Component is more focused on UI and orchestration

## Test Results
- ✅ All 11 E2E tests passing
- ✅ TypeScript compilation successful
- ✅ No breaking changes to existing functionality

## Impact
- **Maintainability**: State transitions and scoring logic easier to understand
- **Testability**: Can now unit test reducer and scoring functions
- **Separation of concerns**: Business logic separated from UI
- **Code organization**: Related logic grouped in dedicated modules
