+++ web/app/play/page.tsx
@@ -13,18 +13,23 @@ import {
   saveResult,
   appendReveal,
   clearReveals,
-  type ResultSummary,
-  type QuestionRecord,
-  type ScoreBreakdown,
   type Outcome,
 } from '@/src/lib/resultStorage';
-import { msToSeconds } from '@/src/lib/timeUtils';
-import type { Question, RoundsStartResponse, RoundsNextResponse, Reveal } from '@/src/features/quiz/api/types';
+import { computePoints, rollupTally, composeSummary } from '@/src/lib/scoring';
+import type { RoundsStartResponse, RoundsNextResponse } from '@/src/features/quiz/api/types';
 import { start, next } from '@/src/features/quiz/datasource';
 import { waitMockReady } from '@/src/lib/waitMockReady';
 import { recordMetricsEvent } from '@/src/lib/metrics/metricsClient';
 import { enrichReveal, toQuestionRecord } from '@/src/features/quiz/lib/reveal';
 import { mark, measure } from '@/src/lib/perfMarks';
+import {
+  playReducer,
+  createInitialState,
+  QUESTION_TIME_LIMIT_MS,
+  TIMEOUT_CHOICE_ID,
+  SKIP_CHOICE_ID,
+  type PlayAction,
+} from '@/src/features/quiz/playReducer';
 
 /**
  * Play page with reveal phase (FE-04 Sub DoD) — refactored with useReducer
@@ -34,181 +39,8 @@ import { mark, measure } from '@/src/lib/perfMarks';
  * - Smaller, testable helpers
  */
 
-type ProgressInfo = { index: number; total: number };
-
-const QUESTION_TIME_LIMIT_MS = 15_000;
-const TIMEOUT_CHOICE_ID = '__timeout__';
-const SKIP_CHOICE_ID = '__skip__';
-
-type State = {
-  token?: string;
-  question?: Question;
-  progress?: ProgressInfo;
-  loading: boolean;
-  error?: string;
-  selectedId?: string;
-  beganAt?: number; // ms timestamp for current question
-  startedAt?: string; // ISO string for run start
-  started: boolean;
-  phase: 'question' | 'reveal';
-  queuedNext?: RoundsNextResponse;
-  currentReveal?: Reveal;
-  deadline?: number;
-  remainingMs: number;
-  tally: ScoreBreakdown;
-  history: QuestionRecord[];
-};
-
 const AUTO_START = process.env.NEXT_PUBLIC_PLAY_AUTOSTART !== '0';
 const IS_MOCK = typeof window !== 'undefined' && process.env.NEXT_PUBLIC_API_MOCK === '1';
-
-// ————————————————————————————————————————————————————————
-// Reducer
-// ————————————————————————————————————————————————————————
-
-type Action =
-  | { type: 'BOOTING' }
-  | {
-      type: 'STARTED';
-      payload: {
-        token: string;
-        question?: Question;
-        progress?: ProgressInfo;
-        beganAt: number;
-        startedAt: string;
-        currentReveal?: Reveal;
-      };
-    }
-  | { type: 'ERROR'; error: string }
-  | { type: 'SELECT'; id: string }
-  | { type: 'ENTER_REVEAL'; reveal?: Reveal }
-  | { type: 'QUEUE_NEXT'; next: RoundsNextResponse; reveal?: Reveal }
-  | { type: 'ADVANCE'; next: RoundsNextResponse }
-  | { type: 'TICK'; remainingMs: number }
-  | { type: 'APPLY_RESULT'; payload: { tally: ScoreBreakdown; history: QuestionRecord[] } };
-
-const EMPTY_TALLY: ScoreBreakdown = { correct: 0, wrong: 0, timeout: 0, skip: 0, points: 0 };
-
-const initialState: State = {
-  loading: AUTO_START,
-  started: AUTO_START,
-  phase: 'question',
-  remainingMs: QUESTION_TIME_LIMIT_MS,
-  tally: { ...EMPTY_TALLY },
-  history: [],
-};
-
-function reducer(state: State, action: Action): State {
-  switch (action.type) {
-    case 'BOOTING':
-      return { ...state, loading: true, error: undefined, started: true };
-
-    case 'STARTED': {
-      const { token, question, progress, beganAt, startedAt, currentReveal } = action.payload;
-      return {
-        token,
-        question,
-        progress,
-        loading: false,
-        error: undefined,
-        selectedId: undefined,
-        beganAt,
-        startedAt,
-        started: true,
-        phase: 'question',
-        queuedNext: undefined,
-        currentReveal,
-        deadline: beganAt + QUESTION_TIME_LIMIT_MS,
-        remainingMs: QUESTION_TIME_LIMIT_MS,
-        tally: { ...EMPTY_TALLY },
-        history: [],
-      };
-    }
-
-    case 'ERROR':
-      return { ...state, loading: false, error: action.error };
-
-    case 'SELECT':
-      return { ...state, selectedId: action.id };
-
-    case 'ENTER_REVEAL':
-      return { ...state, phase: 'reveal', loading: false, error: undefined, currentReveal: action.reveal };
-
-    case 'QUEUE_NEXT': {
-      return { ...state, queuedNext: action.next, currentReveal: action.reveal ?? state.currentReveal };
-    }
-
-    case 'ADVANCE': {
-      const qn = action.next;
-      // NOTE: caller must handle finished-case navigation before dispatching ADVANCE
-      const nextProgress: ProgressInfo | undefined =
-        qn.round?.progress ?? (state.progress ? { index: state.progress.index + 1, total: state.progress.total } : undefined);
-      return {
-        ...state,
-        loading: false,
-        error: undefined,
-        token: qn.round?.token ?? state.token,
-        question: qn.question!,
-        progress: nextProgress,
-        selectedId: undefined,
-        beganAt: performance.now(),
-        phase: 'question',
-        queuedNext: undefined,
-        currentReveal: undefined,
-        deadline: performance.now() + QUESTION_TIME_LIMIT_MS,
-        remainingMs: QUESTION_TIME_LIMIT_MS,
-      };
-    }
-
-    case 'TICK':
-      return { ...state, remainingMs: Math.max(0, Math.round(action.remainingMs)) };
-
-    case 'APPLY_RESULT':
-      return { ...state, tally: action.payload.tally, history: action.payload.history };
-
-    default:
-      return state;
-  }
-}
-
-// ————————————————————————————————————————————————————————
-// Helpers
-// ————————————————————————————————————————————————————————
-
-function computePoints(remainingMs: number): number {
-  return 100 + msToSeconds(remainingMs) * 5;
-}
-
-function rollupTally(prev: ScoreBreakdown, outcome: Outcome, pointsDelta: number): ScoreBreakdown {
-  return {
-    correct: prev.correct + (outcome === 'correct' ? 1 : 0),
-    wrong: prev.wrong + (outcome === 'wrong' ? 1 : 0),
-    timeout: prev.timeout + (outcome === 'timeout' ? 1 : 0),
-    skip: prev.skip + (outcome === 'skip' ? 1 : 0),
-    points: prev.points + pointsDelta,
-  };
-}
-
-function composeSummary(input: {
-  tally: ScoreBreakdown;
-  history: QuestionRecord[];
-  total: number;
-  startedAt?: string;
-  finishedAt: string;
-}): ResultSummary {
-  return {
-    answeredCount: input.history.length,
-    total: input.total,
-    score: input.tally,
-    questions: input.history,
-    startedAt: input.startedAt,
-    finishedAt: input.finishedAt,
-  };
-}
-
-// ————————————————————————————————————————————————————————
-// Component
-// ————————————————————————————————————————————————————————
 
 export default function PlayPage() {
   const router = useRouter();
@@ -216,7 +48,7 @@ export default function PlayPage() {
   const isMountedRef = React.useRef(true);
   React.useEffect(() => () => { isMountedRef.current = false; }, []);
 
-  const [s, dispatch] = React.useReducer(reducer, initialState);
+  const [s, dispatch] = React.useReducer(playReducer, createInitialState(AUTO_START));
   const {
     phase,
     token,
@@ -234,7 +66,7 @@ export default function PlayPage() {
     loading,
   } = s;
 
-  const safeDispatch = React.useCallback((a: Action) => {
+  const safeDispatch = React.useCallback((a: PlayAction) => {
     if (isMountedRef.current) dispatch(a);
   }, []);
